% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.SGD}
\alias{plot.SGD}
\title{Plot method for SGD object}
\usage{
\method{plot}{SGD}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{SGD}}

\item{y}{does nothing}

\item{...}{does nothing}
}
\description{
plot the value of \code{h} across iterations.
}

